step2: Expression 클래스 리팩터링 (feedback)
