--- src/main/kotlin/calculator/Calculator.kt
@@ -1,9 +1,0 @@
-package calculator
-
-class Calculator {
-    fun calculate(expression: String): Double {
-        validateExpression(expression)
-
-        return Expression.calculate(expression)
-    }
-}
