+++ src/main/kotlin/calculator/Expression.kt
@@ -1,28 +1,18 @@
 package calculator
 
-import calculator.Number.Companion.ZERO
+class Expression(
+    private val value: String,
+) {
+    init {
+        require(value.isNotBlank()) { "식이 널이거나, 비어있거나, 공백 문자이면 안 됩니다." }
+    }
 
-object Expression {
-    private const val WHITE_SPACE = " "
+    fun calculate(): Double {
+        val expressions = Expressions(value.split(TOKEN_DELIMITER))
+        return expressions.calculate()
+    }
 
-    fun calculate(expression: String): Double {
-        val expressions = expression.split(WHITE_SPACE)
-        var currentOperator: Operator? = null
-        var result = ZERO
-
-        expressions.forEach { token ->
-            val value = token.toDoubleOrNull()
-
-            if (value == null) {
-                currentOperator = Operator.getSymbol(token)
-                return@forEach
-            }
-            if (currentOperator == null) {
-                result = value
-                return@forEach
-            }
-            result = checkNotNull(currentOperator) { "연산자가 설정되지 않았습니다. operator = $token" }.apply(result, value)
-        }
-        return result
+    companion object {
+        private const val TOKEN_DELIMITER = " "
     }
 }
