+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -1,0 +1,7 @@
+package calculator
+
+class StringCalculator {
+    fun calculate(expression: String): Double {
+        return Expression(expression).calculate()
+    }
+}
